move more calls to v2
